requirements.txt
facebook-business
requests_oauthlib
pandas
pandas_gbq
google-cloud-bigquery
google-auth
requests
numpy
